+++ packages/compat/uni-builder/src/rspack/index.ts
@@ -4,16 +4,18 @@ import type {
   RsbuildPlugin,
   RsbuildInstance,
 } from '@rsbuild/core';
-import type { RsbuildProvider } from '@rsbuild/shared';
+import type { RsbuildProvider, RsbuildTarget } from '@rsbuild/shared';
 import type { UniBuilderRspackConfig } from '../types';
 import type { CreateRspackBuilderOptions } from '../types';
 import { parseCommonConfig } from '../shared/parseCommonConfig';
 import { pluginStyledComponents } from '@rsbuild/plugin-styled-components';
+import { withDefaultConfig } from './defaults';
 
 export async function parseConfig(
   uniBuilderConfig: UniBuilderRspackConfig,
   cwd: string,
   frameworkConfigPath?: string,
+  target?: RsbuildTarget | RsbuildTarget[],
 ): Promise<{
   rsbuildConfig: RsbuildConfig;
   rsbuildPlugins: RsbuildPlugin[];
@@ -22,6 +24,7 @@ export async function parseConfig(
     uniBuilderConfig,
     cwd,
     frameworkConfigPath,
+    target,
   );
 
   if (uniBuilderConfig.output?.enableAssetManifest) {
@@ -31,7 +34,11 @@ export async function parseConfig(
 
   if (uniBuilderConfig.tools?.babel) {
     const { pluginBabel } = await import('@rsbuild/plugin-babel');
-    rsbuildPlugins.push(pluginBabel(uniBuilderConfig.tools?.babel));
+    rsbuildPlugins.push(
+      pluginBabel({
+        babelLoaderOptions: uniBuilderConfig.tools?.babel,
+      }),
+    );
   }
 
   rsbuildPlugins.push(
@@ -47,13 +54,18 @@ export async function parseConfig(
 export async function createRspackBuilder(
   options: CreateRspackBuilderOptions,
 ): Promise<RsbuildInstance<RsbuildProvider>> {
+  const { cwd = process.cwd(), target = 'web' } = options;
+
   const { rsbuildConfig, rsbuildPlugins } = await parseConfig(
-    options.config,
-    options.cwd,
+    withDefaultConfig(options.config),
+    cwd,
     options.frameworkConfigPath,
+    target,
   );
   const rsbuild = await createRsbuild({
+    cwd,
     rsbuildConfig,
+    target,
   });
 
   rsbuild.addPlugins(rsbuildPlugins);