+++ packages/compat/uni-builder/src/webpack/index.ts
@@ -4,6 +4,7 @@ import {
   type RsbuildPlugin,
   type RsbuildInstance,
 } from '@rsbuild/core';
+import type { RsbuildTarget } from '@rsbuild/shared';
 import type { UniBuilderWebpackConfig } from '../types';
 import type { CreateWebpackBuilderOptions } from '../types';
 import { parseCommonConfig } from '../shared/parseCommonConfig';
@@ -11,11 +12,13 @@ import { pluginModuleScopes } from './plugins/moduleScopes';
 import { pluginStyledComponents } from './plugins/styledComponents';
 import { pluginBabel } from './plugins/babel';
 import { pluginReact } from './plugins/react';
+import { withDefaultConfig } from './defaults';
 
 export async function parseConfig(
   uniBuilderConfig: UniBuilderWebpackConfig,
   cwd: string,
   frameworkConfigPath?: string,
+  target?: RsbuildTarget | RsbuildTarget[],
 ): Promise<{
   rsbuildConfig: RsbuildConfig;
   rsbuildPlugins: RsbuildPlugin[];
@@ -24,9 +27,14 @@ export async function parseConfig(
     uniBuilderConfig,
     cwd,
     frameworkConfigPath,
+    target,
   );
 
-  rsbuildPlugins.push(pluginBabel(uniBuilderConfig.tools?.babel));
+  rsbuildPlugins.push(
+    pluginBabel({
+      babelLoaderOptions: uniBuilderConfig.tools?.babel,
+    }),
+  );
   rsbuildPlugins.push(pluginReact());
 
   if (uniBuilderConfig.tools?.tsLoader) {
@@ -69,15 +77,21 @@ export async function parseConfig(
 export async function createWebpackBuilder(
   options: CreateWebpackBuilderOptions,
 ): Promise<RsbuildInstance> {
+  const { cwd = process.cwd(), target = 'web' } = options;
+
   const { rsbuildConfig, rsbuildPlugins } = await parseConfig(
-    options.config,
-    options.cwd,
+    withDefaultConfig(options.config),
+    cwd,
     options.frameworkConfigPath,
+    target,
   );
+
   const { webpackProvider } = await import('@rsbuild/webpack');
   const rsbuild = await createRsbuild({
     rsbuildConfig,
     provider: webpackProvider,
+    target,
+    cwd,
   });
 
   rsbuild.addPlugins([