+++ packages/plugin-babel/compiled/babel-loader/583.index.js
@@ -3,11 +3,11 @@ exports.id = 583;
 exports.ids = [583];
 exports.modules = {
 
-/***/ 2463:
+/***/ 463:
 /***/ ((module, __unused_webpack_exports, __webpack_require__) => {
 
 
-const { sep: DEFAULT_SEPARATOR } = __webpack_require__(1017)
+const { sep: DEFAULT_SEPARATOR } = __webpack_require__(17)
 
 const determineSeparator = paths => {
   for (const path of paths) {
@@ -43,7 +43,7 @@ module.exports = function commonPathPrefix (paths, sep = determineSeparator(path
 
 /***/ }),
 
-/***/ 9583:
+/***/ 583:
 /***/ ((__unused_webpack___webpack_module__, __webpack_exports__, __webpack_require__) => {
 
 // ESM COMPAT FLAG
@@ -55,15 +55,15 @@ __webpack_require__.d(__webpack_exports__, {
 });
 
 // EXTERNAL MODULE: external "node:process"
-var external_node_process_ = __webpack_require__(7742);
+var external_node_process_ = __webpack_require__(742);
 // EXTERNAL MODULE: external "node:path"
-var external_node_path_ = __webpack_require__(9411);
+var external_node_path_ = __webpack_require__(411);
 // EXTERNAL MODULE: external "node:fs"
-var external_node_fs_ = __webpack_require__(7561);
+var external_node_fs_ = __webpack_require__(561);
 // EXTERNAL MODULE: ../../node_modules/.pnpm/common-path-prefix@3.0.0/node_modules/common-path-prefix/index.js
-var common_path_prefix = __webpack_require__(2463);
+var common_path_prefix = __webpack_require__(463);
 // EXTERNAL MODULE: external "node:url"
-var external_node_url_ = __webpack_require__(1041);
+var external_node_url_ = __webpack_require__(41);
 ;// CONCATENATED MODULE: ../../node_modules/.pnpm/yocto-queue@1.0.0/node_modules/yocto-queue/index.js
 /*
 How it works: